Made major changes, most, if not all math is now integer based instead of floating point, divide stats by 100 if you want to use old values (for instance, 100.00 Scrambler HP is now 10,000). Still searching for more issues

+++ Dungeon Scramblers/Assets/Scripts/Hanson's/Dummy.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Dummy : AbstractPlayer
 {
-    public override void Damage(float damageTaken)
+    public override void Damage(int damageTaken)
     {
         Debug.Log(gameObject + " Hit For " + damageTaken + " Damage.");
         GameManager.ManagerInstance.DistributeExperience(10.0f);
